Implement an O(h^2) derivative
This patch adds a memory-friendly routine to calculate first
derivative with an error of O(h^2). It is assumed that the
interval of the independent variable is equal to one, so for
real-life data the calculated derivative will have to be divided
by the value of h.

Note that all calculations are being done in integers, and the
first argument has to be signed.