Implement an O(h^4) derivative
A more precise routine is added to compute the first derivative of
the signal. Just like the O(h^2) version, this new routine is
memory-friendly, assumes an array of signed integers in the first
argument, and thinks that the sampling interval is equal to one.
For the latter reason, division by the value of h is required for
the caller, if the caller wants to get properly scaled data.

See the header file for the macro to switch between O(h^2) and
O(h^4). The more precise version is now the default one.

Also, this patch changes the return type for the derivative
calculation function to void. This means that no sanity check on
the aruments is being done, and therefore this check is caller's
responsibility.